Fix compiler warning

+++ gtkboard.c
@@ -16,7 +16,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
- * $Id: gtkboard.c,v 1.325 2015/06/23 21:49:18 plm Exp $
+ * $Id: gtkboard.c,v 1.326 2015/06/26 06:39:38 mdpetch Exp $
  */
 
 /*! \file gtkboard.c
@@ -2178,7 +2178,7 @@ score_changed(GtkAdjustment * adj, BoardData * bd)
 
     if ((bd->match_to != nMatchLen) && (adj == bd->amatch)) {
         /* reset limits for scores if match length is changed */
-        gfloat upper = (gfloat) (nMatchLen == 0) ? 32767 : nMatchLen - 1;
+        gdouble upper = (gdouble) ((nMatchLen == 0) ? 32767 : nMatchLen - 1);
         gtk_adjustment_set_upper(bd->ascore0, upper);
         gtk_adjustment_set_upper(bd->ascore1, upper);
     }
